refactor: model -> entity 로 패키지 네이밍 수정
