+++ src/main/java/roomescape/domain/reservation/dao/JdbcReservationRepository.java
@@ -15,7 +15,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.reservation.exception.ReservationException;
-import roomescape.domain.reservation.model.Reservation;
+import roomescape.domain.reservation.entity.Reservation;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
